Extract simplified AccordionItem component
